+++ src/clojure/halti_server/api/instances/lifecycle.clj
@@ -26,6 +26,7 @@
   {:instance_id (uuid)
    :created (t/now)
    :containers []
+   :capabilities []
    :last_heartbeat (t/now)})